+++ tests/testthat/test-enw_expectation.R
@@ -13,6 +13,7 @@ test_that("enw_expectation produces the expected default model", {
     exp$inits(c(exp$data, obs$data), exp$priors)(),
     c(
       "expr_beta", "expr_beta_sd", "expr_lelatent_int", "expr_r_int",
+      "expr_gt_mean", "expr_gt_sd", "expl_lrd_mean", "expl_lrd_sd",
       "expl_beta", "expl_beta_sd",
       "expr_arima_pacf", "expr_arima_theta", "expr_arima_sigma",
       "expr_gp_rho", "expr_gp_alpha",
@@ -117,7 +118,9 @@ test_that(
       ),
       list(
         expr_beta = 3L, expr_beta_sd = 1L, expr_lelatent_int = c(2L, 3L),
-        expr_r_int = 1L, expl_beta = 6L, expl_beta_sd = 3L,
+        expr_r_int = 1L, expr_gt_mean = NULL, expr_gt_sd = NULL,
+        expl_lrd_mean = NULL, expl_lrd_sd = NULL,
+        expl_beta = 6L, expl_beta_sd = 3L,
         expr_arima_pacf = NULL, expr_arima_theta = NULL,
         expr_arima_sigma = NULL,
         expr_gp_rho = NULL, expr_gp_alpha = NULL,